Checking in to test on another system.
M    src/GenerateProblem.cpp
M    src/main.cpp
M    src/mytimer.cpp
M    src/dot.cpp
M    src/ExchangeHalo.cpp
M    src/CG.cpp
M    src/Makefile
M    src/ReportResults.cpp
M    src/Geometry.hpp
M    src/ComputeResidual.cpp
M    src/OptimizeMatrix.cpp
M    src/SparseMatrix.hpp


git-svn-id: svn+ssh://software.sandia.gov/svn/hpcg/trunk@33 5e06a57e-6b39-4354-92a0-249a7d0c5410

+++ src/Geometry.hpp
@@ -11,8 +11,8 @@
 #ifndef GEOMETRY_HPP
 #define GEOMETRY_HPP
 
-typedef long long local_int_t;
-typedef long long global_int_t;
+typedef int local_int_t;
+typedef int global_int_t;
 
 // This is a data structure to contain all processor navigation information
